Test case where LDOS fails to allocate when there should be space.
I've put this test together to demonstrate the problem I'm having. In short I am stopping the music which frees the bank. I am then loading a blob with a different bank that is smaller 225k versus 248k. I would expect this to fit without issue but it's asserting.
If I switch to a bank file that's significantly smaller 205k then it does work. All the memory reports seem to suggest there is room, it just won;t load?